+++ lib/src/application/bus.dart
@@ -4,7 +4,6 @@ class ArcheBus extends TypeProvider {
   ArcheBus._();
   factory ArcheBus() => singleton.provideof<ArcheBus>(instance: ArcheBus._());
   static ArcheBus get bus => ArcheBus();
-  static busof<T>() => bus.of();
   static ArcheConfig get config => bus.of();
   static ArcheLogger get logger => bus.of();
   ArcheConfig get getConfig => of();