+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+ - Imporve the `NavigationView`, support the vertical view
+ - `tryGet` method in `KVIO`
+ - Some Small Update and BugFix
+
 ## 0.0.3
 
  - Optimize the project structure